+++ test/feature_filters/time_window_filter_test.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:azure_app_config/src/feature_filter.dart';
+import 'package:azure_app_config/src/feature_filters/time_window_filter.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -8,34 +8,7 @@ void main() {
   final current = DateTime.utc(2022, 09, 21, 14, 22, 50);
   final after = DateTime.utc(2022, 10, 21, 14, 22, 50);
 
-  var percentageFilter = TimeWindow(clock: current);
-
-  test('''
-Percentage filter should return the same value for the same user string''', () {
-    final params = <String, dynamic>{
-      'Audience': {
-        'DefaultRolloutPercentage': 25,
-      },
-    };
-
-    final filter = FeatureFilter.percentage();
-
-    var actual = filter.evaluate(
-      params,
-      const FeatureFilterSettings(user: 'test.user1@company.com'),
-      'feature-a',
-    );
-
-    expect(actual, true);
-
-    actual = filter.evaluate(
-      params,
-      const FeatureFilterSettings(user: 'test.user2@company.com'),
-      'feature-a',
-    );
-
-    expect(actual, false);
-  });
+  var timewindowFilter = TimeWindowFilter(clock: current);
 
   test('Time Filter should return good result based on current time and input',
       () {
@@ -44,7 +17,7 @@ Percentage filter should return the same value for the same user string''', () {
       'Start': HttpDate.format(before),
     };
 
-    var actual = percentageFilter.evaluate(params, null, '');
+    var actual = timewindowFilter.evaluate(params, '');
     var expected = true;
 
     expect(actual, expected);
@@ -54,42 +27,42 @@ Percentage filter should return the same value for the same user string''', () {
       'End': HttpDate.format(after),
     };
 
-    actual = percentageFilter.evaluate(params, null, '');
+    actual = timewindowFilter.evaluate(params, '');
     expected = true;
 
     expect(actual, expected);
 
     // START AND END
-    percentageFilter = TimeWindow(clock: before);
+    timewindowFilter = TimeWindowFilter(clock: before);
     params = {
       'Start': HttpDate.format(current),
       'End': HttpDate.format(after),
     };
 
-    actual = percentageFilter.evaluate(params, null, '');
+    actual = timewindowFilter.evaluate(params, '');
     expected = false;
     expect(actual, expected);
 
-    percentageFilter = TimeWindow(clock: after);
+    timewindowFilter = TimeWindowFilter(clock: after);
     params = {
       'Start': HttpDate.format(before),
       'End': HttpDate.format(current),
     };
 
-    actual = percentageFilter.evaluate(params, null, '');
+    actual = timewindowFilter.evaluate(params, '');
     expected = false;
     expect(actual, expected);
 
     // Invalid Inputs should return true
     params = <String, dynamic>{};
-    actual = percentageFilter.evaluate(params, null, '');
+    actual = timewindowFilter.evaluate(params, '');
     expected = true;
 
     expect(actual, expected);
 
     params = <String, dynamic>{'Start': 23};
 
-    actual = percentageFilter.evaluate(params, null, '');
+    actual = timewindowFilter.evaluate(params, '');
     expected = true;
 
     expect(actual, expected);